+++ src/libraries/BCAL/DBCALUnifiedHit_factory.cc
@@ -40,7 +40,7 @@ jerror_t DBCALUnifiedHit_factory::init(void)
     return NOERROR;
 }
 
-jerror_t DBCALUnifiedHit_factory::brun(jana::JEventLoop *eventLoop, int runnumber) {
+jerror_t DBCALUnifiedHit_factory::brun(jana::JEventLoop *eventLoop, int32_t runnumber) {
 
     if (USE_TDC){
         //get timewalk corrections from CCDB
@@ -95,7 +95,7 @@ jerror_t DBCALUnifiedHit_factory::brun(jana::JEventLoop *eventLoop, int runnumbe
 //----------------
 // evnt
 //----------------
-jerror_t DBCALUnifiedHit_factory::evnt(JEventLoop *loop, int eventnumber) {
+jerror_t DBCALUnifiedHit_factory::evnt(JEventLoop *loop, uint64_t eventnumber) {
     vector<const DBCALHit*> hits;
     vector<const DBCALTDCHit*> tdc_hits;
     loop->Get(hits);
@@ -191,8 +191,8 @@ jerror_t DBCALUnifiedHit_factory::evnt(JEventLoop *loop, int eventnumber) {
                         t_diff=tdc_adc_diff;
                     }
                     if (VERBOSE>5) {
-                        printf("DBCALUnifiedHit_factory  event %5i (%2i,%i,%i,%i) TDC %i %6.1f  ADC %6.1f diff %6.1f    best %2i %6.1f\n",
-                                eventnumber, module, layer, sector, chan.end, i, tdc_hit->t, hit->t, tdc_adc_diff, goodTDCindex, t_diff);
+                        printf("DBCALUnifiedHit_factory  event %5llu (%2i,%i,%i,%i) TDC %i %6.1f  ADC %6.1f diff %6.1f    best %2i %6.1f\n",
+                                (unsigned long long int)eventnumber, module, layer, sector, chan.end, i, tdc_hit->t, hit->t, tdc_adc_diff, goodTDCindex, t_diff);
                     }
                 }
                 t_TDC = tdc_hits[goodTDCindex]->t;
