+++ src/libraries/CDC/DCDCHit_factory.cc
@@ -49,7 +49,7 @@ jerror_t DCDCHit_factory::init(void)
 //------------------
 // brun
 //------------------
-jerror_t DCDCHit_factory::brun(jana::JEventLoop *eventLoop, int runnumber)
+jerror_t DCDCHit_factory::brun(jana::JEventLoop *eventLoop, int32_t runnumber)
 {
     // Only print messages for one thread whenever run number change
     static pthread_mutex_t print_mutex = PTHREAD_MUTEX_INITIALIZER;
@@ -163,7 +163,7 @@ jerror_t DCDCHit_factory::brun(jana::JEventLoop *eventLoop, int runnumber)
 //------------------
 // evnt
 //------------------
-jerror_t DCDCHit_factory::evnt(JEventLoop *loop, int eventnumber)
+jerror_t DCDCHit_factory::evnt(JEventLoop *loop, uint64_t eventnumber)
 {
     /// Generate DCDCHit object for each DCDCDigiHit object.
     /// This is where the first set of calibration constants
@@ -276,7 +276,7 @@ jerror_t DCDCHit_factory::fini(void)
 //------------------
 // CalcNstraws
 //------------------
-void DCDCHit_factory::CalcNstraws(jana::JEventLoop *eventLoop, int runnumber, vector<unsigned int> &Nstraws)
+void DCDCHit_factory::CalcNstraws(jana::JEventLoop *eventLoop, int32_t runnumber, vector<unsigned int> &Nstraws)
 {
     DGeometry *dgeom;
     vector<vector<DCDCWire *> >cdcwires;
