+++ src/libraries/PID/DBeamPhoton_factory_TRUTH.cc
@@ -23,7 +23,7 @@ jerror_t DBeamPhoton_factory_TRUTH::init(void)
 //------------------
 // brun
 //------------------
-jerror_t DBeamPhoton_factory_TRUTH::brun(jana::JEventLoop *locEventLoop, int runnumber)
+jerror_t DBeamPhoton_factory_TRUTH::brun(jana::JEventLoop *locEventLoop, int32_t runnumber)
 {
 	DApplication* dapp = dynamic_cast<DApplication*>(locEventLoop->GetJApplication());
 	DGeometry* locGeometry = dapp->GetDGeometry(locEventLoop->GetJEvent().GetRunNumber());
@@ -36,7 +36,7 @@ jerror_t DBeamPhoton_factory_TRUTH::brun(jana::JEventLoop *locEventLoop, int run
 //------------------
 // evnt
 //------------------
-jerror_t DBeamPhoton_factory_TRUTH::evnt(jana::JEventLoop *locEventLoop, int eventnumber)
+jerror_t DBeamPhoton_factory_TRUTH::evnt(jana::JEventLoop *locEventLoop, uint64_t eventnumber)
 {
 	DVector3 pos(0.0, 0.0, dTargetCenterZ);
