+++ src/plugins/Analysis/rho_p_hists/DEventProcessor_rho_p_hists.h
@@ -43,7 +43,7 @@ class DEventProcessor_rho_p_hists:public JEventProcessor{
 
 	private:
 		jerror_t init(void);	///< Invoked via DEventProcessor virtual method
-		jerror_t evnt(JEventLoop *loop, int eventnumber);	///< Invoked via DEventProcessor virtual method
+		jerror_t evnt(JEventLoop *loop, uint64_t eventnumber);	///< Invoked via DEventProcessor virtual method
 		jerror_t erun(void);					///< Invoked via DEventProcessor virtual method
 		jerror_t fini(void);					///< Invoked via DEventProcessor virtual method
