+++ src/plugins/monitoring/L3_online/JEventProcessor_L3_online.cc
@@ -84,7 +84,7 @@ jerror_t JEventProcessor_L3_online::init(void) {
 //----------------------------------------------------------------------------------
 
 
-jerror_t JEventProcessor_L3_online::brun(JEventLoop *eventLoop, int runnumber) {
+jerror_t JEventProcessor_L3_online::brun(JEventLoop *eventLoop, int32_t runnumber) {
   // This is called whenever the run number changes
   return NOERROR;
 }
@@ -93,7 +93,7 @@ jerror_t JEventProcessor_L3_online::brun(JEventLoop *eventLoop, int runnumber) {
 //----------------------------------------------------------------------------------
 
 
-jerror_t JEventProcessor_L3_online::evnt(JEventLoop *eventLoop, int eventnumber) {
+jerror_t JEventProcessor_L3_online::evnt(JEventLoop *eventLoop, uint64_t eventnumber) {
 
   // This is called for every event. Use of common resources like writing
   // to a file or filling a histogram should be mutex protected. Using
