+++ src/plugins/monitoring/PS_online/JEventProcessor_PS_online.cc
@@ -180,7 +180,7 @@ jerror_t JEventProcessor_PS_online::init(void) {
 //----------------------------------------------------------------------------------
 
 
-jerror_t JEventProcessor_PS_online::brun(JEventLoop *eventLoop, int runnumber) {
+jerror_t JEventProcessor_PS_online::brun(JEventLoop *eventLoop, int32_t runnumber) {
   // This is called whenever the run number changes
   // extract the PS geometry
   vector<const DPSGeometry*> psGeomVect;
@@ -216,7 +216,7 @@ jerror_t JEventProcessor_PS_online::brun(JEventLoop *eventLoop, int runnumber) {
 //----------------------------------------------------------------------------------
 
 
-jerror_t JEventProcessor_PS_online::evnt(JEventLoop *eventLoop, int eventnumber) {
+jerror_t JEventProcessor_PS_online::evnt(JEventLoop *eventLoop, uint64_t eventnumber) {
   // This is called for every event. Use of common resources like writing
   // to a file or filling a histogram should be mutex protected. Using
   // loop-Get(...) to get reconstructed objects (and thereby activating the
