+++ src/programs/Analysis/hd_eventfilter/MyProcessor.cc
@@ -69,7 +69,7 @@ jerror_t MyProcessor::init(void)
 //------------------------------------------------------------------
 // evnt   -Fill histograms here
 //------------------------------------------------------------------
-jerror_t MyProcessor::evnt(JEventLoop *loop, int eventnumber)
+jerror_t MyProcessor::evnt(JEventLoop *loop, uint64_t eventnumber)
 {
    JEvent& event = loop->GetJEvent();
    JEventSource *source = event.GetJEventSource();
