+++ src/programs/Analysis/hd_example/MyProcessor.cc
@@ -34,7 +34,7 @@ jerror_t MyProcessor::init(void)
 //------------------------------------------------------------------
 // evnt   -Fill histograms here
 //------------------------------------------------------------------
-jerror_t MyProcessor::evnt(JEventLoop *eventLoop, int eventnumber)
+jerror_t MyProcessor::evnt(JEventLoop *eventLoop, uint64_t eventnumber)
 {
 	vector<const DCDCHit*> cdchits;
 	vector<const DFCALHit*> fcalhits;
